pyproject.toml
[tool.poetry]
name = "testfastapiproj"
version = "0.1.0"
description = ""
authors = ["SV <svoloshyn614@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
aiohttp = "^3.10.8"
uvicorn = "^0.31.0"
python-dotenv = "^1.0.1"
aioboto3 = "^13.1.1"
aiofiles = "^24.1.0"
flake8 = "^7.1.1"
pydantic = "^2.9.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
